+++ fixtures/function-source/src/SpecialFunctionsImportedParent.sol
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.13;
+
+contract SpecialFunctionsImportedParent {
+    function importedParentFunction() internal pure returns (uint256) {
+        return 3;
+    }
+}